backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
httpx==0.27.0
python-dotenv==1.0.1
numpy==1.26.4
scipy==1.11.4
pydantic==2.6.4
pydantic-settings==2.1.0
google-generativeai==0.6.0
tenacity==8.2.3
orjson==3.10.6
beautifulsoup4==4.12.3
